refactor(assistant-runtime): extract hosted group tool context
